+++ src/test/java/org/cicirello/search/concurrent/ParallelMultistarterValidator.java
@@ -1,0 +1,59 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2023 Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ *
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.concurrent;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.cicirello.search.problems.OptimizationProblem;
+import org.cicirello.util.Copyable;
+import org.junit.jupiter.api.*;
+
+/** Test validation common to multiple test classes for testing parallel multistarters. */
+public class ParallelMultistarterValidator {
+
+  static class TestObject implements Copyable<TestObject> {
+
+    public TestObject() {}
+
+    @Override
+    public TestObject copy() {
+      return new TestObject();
+    }
+  }
+
+  static class TestProblem implements OptimizationProblem<TestObject> {
+    public double cost(TestObject o) {
+      return 5;
+    }
+
+    public boolean isMinCost(double c) {
+      return false;
+    }
+
+    public double minCost() {
+      return -10000;
+    }
+
+    public double value(TestObject o) {
+      return 5;
+    }
+  }
+}
